+++ packages/backend/src/database/seeds/development/01_initial_user.ts
@@ -175,7 +175,7 @@ export async function seed(knex: Knex): Promise<void> {
             ...SEED_PROJECT,
             organization_id: organizationId,
             dbt_connection: encryptedProjectSettings,
-            dbt_version: SupportedDbtVersions.V1_4,
+            dbt_version: SupportedDbtVersions.V1_7,
             semantic_layer_connection: null,
             created_by_user_uuid: user.user_uuid,
         })
@@ -248,7 +248,7 @@ export async function seed(knex: Knex): Promise<void> {
                 warehouseCatalog: undefined,
                 onWarehouseCatalogChange: () => {},
             },
-            SupportedDbtVersions.V1_4,
+            SupportedDbtVersions.V1_7,
         );
         const explores = await adapter.compileAllExplores({
             userUuid: user.user_uuid,