+++ packages/backend/src/utils/QueryBuilder/queryBuilder.ts
@@ -25,6 +25,7 @@ import {
     ItemsMap,
     MetricFilterRule,
     parseAllReferences,
+    QueryWarning,
     renderFilterRuleSql,
     renderFilterRuleSqlFromField,
     renderTableCalculationFilterRuleSql,
@@ -34,9 +35,10 @@ import {
     WarehouseClient,
     WeekDay,
 } from '@lightdash/common';
-import { wrapSentryTransactionSync } from '../../utils';
+import Logger from '../../logging/logger';
 import {
     assertValidDimensionRequiredAttribute,
+    findMetricInflationWarnings,
     getCustomBinDimensionSql,
     getCustomSqlDimensionSql,
     getDimensionFromFilterTargetId,
@@ -53,6 +55,7 @@ import {
 export type CompiledQuery = {
     query: string;
     fields: ItemsMap;
+    warnings: QueryWarning[];
 };
 
 export type BuildQueryProps = {
@@ -64,15 +67,27 @@ export type BuildQueryProps = {
     timezone: string;
 };
 
-export const buildQuery = ({
-    explore,
-    compiledMetricQuery,
-    warehouseClient,
-    intrinsicUserAttributes,
-    userAttributes = {},
-    timezone,
-}: BuildQueryProps): CompiledQuery =>
-    wrapSentryTransactionSync('QueryBuilder.buildQuery', {}, () => {
+export class MetricQueryBuilder {
+    constructor(private args: BuildQueryProps) {}
+
+    /**
+     * Compiles a database query based on the provided metric query, explores, user attributes, and warehouse-specific configurations.
+     *
+     * This method processes dimensions, metrics, filters, and joins across multiple dataset definitions to generate
+     * a complete SQL query string tailored for the specific warehouse type and environment. Additionally, it ensures
+     * field validation and substitution of user-specific attributes for dynamic query generation.
+     *
+     * @return {CompiledQuery} The compiled query object containing the SQL string and meta information ready for execution.
+     */
+    public compileQuery(): CompiledQuery {
+        const {
+            explore,
+            compiledMetricQuery,
+            warehouseClient,
+            intrinsicUserAttributes,
+            userAttributes = {},
+            timezone,
+        } = this.args;
         const fields = getFieldsFromMetricQuery(compiledMetricQuery, explore);
         const adapterType: SupportedDbtAdapter =
             warehouseClient.getAdapterType();
@@ -557,6 +572,20 @@ export const buildQuery = ({
             filters.tableCalculations,
         );
 
+        let warnings: QueryWarning[] = [];
+        try {
+            warnings = findMetricInflationWarnings({
+                possibleJoins: explore.joinedTables,
+                baseTable: explore.baseTable,
+                joinedTables,
+                metrics: metrics.map((field) =>
+                    getMetricFromId(field, explore, compiledMetricQuery),
+                ),
+            });
+        } catch (e) {
+            Logger.error('Error during metric inflation detection', e);
+        }
+
         const sqlLimit = `LIMIT ${limit}`;
 
         if (
@@ -616,6 +645,7 @@ export const buildQuery = ({
             return {
                 query: [cte, finalQuery, sqlOrderBy, sqlLimit].join('\n'),
                 fields,
+                warnings,
             };
         }
 
@@ -640,8 +670,10 @@ export const buildQuery = ({
         return {
             query: metricQuerySql,
             fields,
+            warnings,
         };
-    });
+    }
+}
 
 type ReferenceObject = { type: DimensionType; sql: string };
 export type ReferenceMap = Record<string, ReferenceObject> | undefined;